+++ src/_pytask/report.py
@@ -1,3 +1,4 @@
+"""This module contains everything related to reports."""
 import math
 import re
 from pathlib import Path
@@ -10,6 +11,8 @@ from _pytask.traceback import remove_internal_traceback_frames_from_exc_info
 
 @attr.s
 class CollectionReport:
+    """A general collection report."""
+
     title = attr.ib(type=str)
     exc_info = attr.ib(type=tuple)
 
@@ -23,6 +26,8 @@ class CollectionReport:
 
 @attr.s
 class CollectionReportTask:
+    """A collection report for a task."""
+
     path = attr.ib(type=Path)
     name = attr.ib(type=str)
     task = attr.ib(default=None)
@@ -46,6 +51,8 @@ class CollectionReportTask:
 
 @attr.s
 class CollectionReportFile:
+    """A collection report for a file."""
+
     path = attr.ib(type=Path)
     exc_info = attr.ib(default=None)
 
@@ -63,26 +70,36 @@ class CollectionReportFile:
 
 @attr.s
 class ResolvingDependenciesReport:
+    """A report for an error while resolving dependencies."""
+
     exc_info = attr.ib()
+
+    @classmethod
+    def from_exception(cls, exc_info):
+        return cls(exc_info)
 
 
 @attr.s
 class ExecutionReport:
+    """A report for an executed task."""
+
     task = attr.ib()
     success = attr.ib(type=bool)
     exc_info = attr.ib(default=None)
+    sections = attr.ib(factory=list)
 
     @classmethod
     def from_task_and_exception(cls, task, exc_info):
         exc_info = remove_internal_traceback_frames_from_exc_info(exc_info)
-        return cls(task, False, exc_info)
+        return cls(task, False, exc_info, task._report_sections)
 
     @classmethod
     def from_task(cls, task):
-        return cls(task, True)
+        return cls(task, True, None, task._report_sections)
 
 
 def format_execute_footer(n_successful, n_failed, duration, terminal_width):
+    """Format the footer of the execution."""
     message = []
     if n_successful:
         message.append(
@@ -93,7 +110,7 @@ def format_execute_footer(n_successful, n_failed, duration, terminal_width):
     message = " " + ", ".join(message) + " "
 
     color = ColorCode.FAILED.value if n_failed else ColorCode.SUCCESS.value
-    message += click.style(f"in {duration} second(s) ", fg=color)
+    message += click.style(f"in {duration}s ", fg=color)
 
     formatted_message = _wrap_string_ignoring_ansi_colors(
         message, color, terminal_width
@@ -103,6 +120,12 @@ def format_execute_footer(n_successful, n_failed, duration, terminal_width):
 
 
 def _wrap_string_ignoring_ansi_colors(message, color, width):
+    """Wrap a string with ANSI colors.
+
+    This wrapper ignores the color codes which will increase the length of the string,
+    but will not show up in the printed string.
+
+    """
     n_characters = width - len(_remove_ansi_colors(message))
     n_left, n_right = math.floor(n_characters / 2), math.ceil(n_characters / 2)
 
@@ -114,5 +137,6 @@ def _wrap_string_ignoring_ansi_colors(message, color, width):
 
 
 def _remove_ansi_colors(string):
+    """Remove ANSI colors from a string."""
     ansi_escape = re.compile(r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])")
     return ansi_escape.sub("", string)